+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [v0.0.2]
+
+- Add AI analysis of found issues on hover
+- Check if weggli and ripgrep are installed
+- Remove Welcome to Greppy from side panel
+- Use greppy instead of greppy.patterns for cog icon
+- Fix: Initialize pattern sets with built-in patterns
+
 ## [v0.0.1]
 
 - Initial release